+++ src/main/java/com/wordonline/matching/auth/service/UserService.java
@@ -10,18 +10,12 @@ import com.wordonline.matching.auth.domain.UserStatus;
 import com.wordonline.matching.auth.dto.UserDetailResponseDto;
 import com.wordonline.matching.auth.dto.UserResponseDto;
 import com.wordonline.matching.auth.repository.UserRepository;
-import com.wordonline.matching.deck.service.DeckInitializer;
-import com.wordonline.matching.deck.service.DeckService;
-import com.wordonline.matching.decoration.service.DecorationInitializer;
-import com.wordonline.matching.magic.service.MagicService;
 import com.wordonline.matching.matching.client.AccountClient;
-import com.wordonline.matching.quest.service.QuestInitializer;
-import com.wordonline.matching.service.LocalizationService;
+import com.wordonline.matching.global.service.LocalizationService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
-import reactor.util.function.Tuples;
 
 @Slf4j
 @Service
